requirements.txt
aiohttp[speedups]
beautifulsoup4
cloudscraper
fastapi==0.104.1
gunicorn
mangum
requests
uvicorn[standard]
sqlalchemy
alembic
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-prometheus